ADDED SIGQUIT SIGPIPE correction deconnexion user non connecté
