+++ code/bonus/Makefile
@@ -6,11 +6,11 @@
 #    By: smlamali <smlamali@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/03/19 20:20:53 by smlamali          #+#    #+#              #
-#    Updated: 2025/03/20 21:05:20 by smlamali         ###   ########.fr        #
+#    Updated: 2025/03/23 17:30:57 by smlamali         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
-SRCS = Bot.cpp main.cpp Actions.cpp
+SRCS = Bot.cpp main.cpp Actions.cpp Utils.cpp
 
 CC = c++
