+++ code/class/Irssi_serv/other/geter.cpp
@@ -6,7 +6,7 @@
 /*   By: smlamali <smlamali@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/24 23:24:28 by yzaoui            #+#    #+#             */
-/*   Updated: 2025/03/17 15:50:43 by smlamali         ###   ########.fr       */
+/*   Updated: 2025/03/23 17:28:18 by smlamali         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
