+++ code/class/Irssi_serv/other/other.cpp
@@ -6,7 +6,7 @@
 /*   By: smlamali <smlamali@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/26 22:14:20 by yzaoui            #+#    #+#             */
-/*   Updated: 2025/03/18 14:09:51 by smlamali         ###   ########.fr       */
+/*   Updated: 2025/03/23 17:28:17 by smlamali         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
