+++ code/utils/ft_utils.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_utils.cpp                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yzaoui <yzaoui@student.42.fr>              +#+  +:+       +#+        */
+/*   By: smlamali <smlamali@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 03:21:56 by yzaoui            #+#    #+#             */
-/*   Updated: 2025/03/02 17:48:57 by yzaoui           ###   ########.fr       */
+/*   Updated: 2025/03/23 17:01:18 by smlamali         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
